update check reward signature

+++ btcserver/internal/api/stake.go
@@ -90,7 +90,7 @@ func (s *Server) SubmitProofOfStake(c *gin.Context) {
 
 	// verify reward receiver signature
 	if err := s.btcClient.CheckRewardAddressSignature(stakeInfo.StakerPublicKey, stakeInfo.RewardReceiver,
-		stakeInfo.ReceiverSignature, int32(stakeInfo.Timestamp),
+		stakeInfo.ReceiverSignature, stakeInfo.Timestamp,
 	); err != nil {
 		respData.Msg = err.Error()
 		JSONResp(c, http.StatusBadRequest, respData)
